Working program of getting a year of JSON data
This works on getting a years worth of JSON data from apod.  Some problems that exist though is that the way python writes the JSON data as a string and doesn't look like JSON at all.  There is a also a minor problem that this program doesn't currently check for when someone decide to start collecting on a leap year day when the year is not a leap year.